+++ goods/urls.py
@@ -1,0 +1,7 @@
+from django.conf.urls import url
+from . import views
+
+urlpatterns = [
+    url(r'^detail/(?P<id>[0-9]+)/$',views.detail, name='detail')
+    #url(r'^', views.index, name='index'),
+]
