+++ main/models.py
@@ -48,34 +48,6 @@ class CartInfo(models.Model):
     isDelete = models.BooleanField(default=False)
 
 
-class GoodInfo(models.Model):
-    gname = models.CharField(max_length=20)
-    gdescrption = models.CharField(max_length=200)
-    dictID = models.ForeignKey('Dict')
-    gunit = models.CharField(max_length=20)
-    gprice = models.CharField(max_length=20)
-    gdetail = HTMLField()
-    gpic = models.ImageField()
-
-
-class Comment(models.Model):
-    goodID = models.ForeignKey('GoodInfo')
-    userID = models.ForeignKey('UserInfo')
-    datetime = models.DateTimeField()
-    content = models.CharField(max_length=100)
-
-
-class ViewInfo(models.Model):
-    ipAddr = models.CharField(max_length=20)
-    goodID = models.ForeignKey('GoodInfo')
-
-
-class Dict(models.Model):
-    title = models.CharField(max_length=20)
-    pid = models.ForeignKey('self', null=True, blank=True)
-    isDelete = models.BooleanField(default=False)
-
-
 class Log(models.Model):
 	record = models.CharField(max_length=1000)
 	datetime = models.DateTimeField()
