+++ templates/base.html
@@ -8,6 +8,8 @@
 	<script type="text/javascript" src="/static/js/jquery-1.12.4.min.js"></script>
 	<script type="text/javascript" src="/static/js/jquery-ui.min.js"></script>
 	<script type="text/javascript" src="/static/js/slide.js"></script>
+
+	{%block javascript%}{%endblock%}
 </head>
 <body>
 	<div class="header_con">
@@ -48,7 +50,7 @@
 		<p>CopyRight © 2016 北京天天生鲜信息技术有限公司 All Rights Reserved</p>
 		<p>电话：010-****888    京ICP备*******8号</p>
 	</div>
-	<script type="text/javascript" src="/static/js/slideshow.js"></script>
+<!-- 	<script type="text/javascript" src="/static/js/slideshow.js"></script>	
 	<script type="text/javascript">
 		BCSlideshow('focuspic');
 		var oFruit = document.getElementById('fruit_more');
@@ -64,6 +66,6 @@
 		oFruit.onclick = function(){
 			window.location.href = 'list.html';
 		}
-	</script>
+	</script> -->
 </body>
 </html>
